+++ src/Pattern/Pattern.php
@@ -3,87 +3,21 @@
   namespace Funivan\PhpTokenizer\Pattern;
 
   use Funivan\PhpTokenizer\Collection;
-  use Funivan\PhpTokenizer\Exception\Exception;
-  use Funivan\PhpTokenizer\QuerySequence\QuerySequence;
+
 
   /**
-   * Apply pattern to collection
+   * @deprecated
    */
-  class Pattern implements PatternCheckerInterface {
+  class Pattern extends PatternMatcher {
 
     /**
-     * @var Collection[]
-     */
-    protected $collections = [];
-
-
-    /**
-     *
-     * @param Collection $collection
+     * @deprecated
+     * @inheritdoc
      */
     public function __construct(Collection $collection) {
-      $this->collections[] = $collection;
+      trigger_error('Deprecated. Use matcher instead', E_USER_DEPRECATED);
+      parent::__construct($collection);
     }
 
-
-    /**
-     * @inheritdoc
-     */
-    public function apply(callable $pattern) {
-
-      # Clear current collections.
-      # We will add new one and iterate over current
-
-      $collections = $this->collections;
-      $this->collections = [];
-
-      foreach ($collections as $collection) {
-
-        $collectionsResult = $this->iterateOverCollections($pattern, $collection);
-
-        foreach ($collectionsResult as $resultCollection) {
-          $this->collections[] = $resultCollection;
-        }
-
-      }
-
-      return $this;
-    }
-
-
-    /**
-     * @inheritdoc
-     */
-    public function getCollections() {
-      return $this->collections;
-    }
-
-
-    /**
-     * @param callable $pattern
-     * @param Collection $collection
-     * @return Collection[]
-     * @throws Exception
-     */
-    protected function iterateOverCollections(callable $pattern, Collection $collection) {
-      $result = [];
-
-      $collection->rewind();
-      foreach ($collection as $index => $token) {
-        $querySequence = new QuerySequence($collection, $index);
-        $patternResult = $pattern($querySequence);
-        if ($patternResult === null) {
-          continue;
-        }
-
-        if (!($patternResult instanceof Collection)) {
-          throw new Exception('Invalid result from pattern callback. Expect Collection. Given:' . gettype($patternResult));
-        }
-
-        $result[] = $patternResult;
-      }
-
-      return $result;
-    }
 
   }