+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All Notable changes to `PhpTokenizer` will be documented in this file
 
 
 ## [UNRELEASED] 
+ - deprecated `Funivan\PhpTokenizer\Pattern\Pattern` use `Funivan\PhpTokenizer\Pattern\PatternMatcher`
  - deprecated `ClassPattern::nameIs`
  - deprecated `ClassPattern::whereName`
  - fixed  `Funivan\PhpTokenizer\Helper::getTokensFromString`. Better detection of token line
@@ -10,6 +11,7 @@ All Notable changes to `PhpTokenizer` will be documented in this file
  - changed `Funivan\PhpTokenizer\Strategy\Section` Strict detection of section. Next token after current must be section start token 
  - added `Funivan\PhpTokenizer\Pattern\Patterns\ClassPattern` `withDocComment`, `withoutDocComment`, `withPossibleDocComment`
  - added `Funivan\PhpTokenizer\Pattern\Patterns\ClassPattern` `withModifier`, `withoutModifier`, `withAnyModifier`
+ - added `Funivan\PhpTokenizer\Pattern\Patterns\ArgumentsPattern`
 
 ## 0.1.1 - 2015-07-28
   - fixed `\Funivan\PhpTokenizer\Strategy\Search::process`